+++ src/GeneratorEngineServiceProvider.php
@@ -7,7 +7,10 @@ use Blutrixx\GeneratorEngine\Commands\MakeUxFromBlueprintCommand;
 
 class GeneratorEngineServiceProvider extends ServiceProvider
 {
-    public function register(): void {}
+    public function register(): void
+    {
+        $this->mergeConfigFrom(__DIR__.'/../config/generator.php', 'generator');
+    }
 
     public function boot(): void
     {
@@ -15,6 +18,9 @@ class GeneratorEngineServiceProvider extends ServiceProvider
             $this->commands([
                 MakeUxFromBlueprintCommand::class,
             ]);
+
+            $this->publishes([__DIR__.'/../config/generator.php' => config_path('generator.php')], 'generator-config');
+            $this->publishes([__DIR__.'/Generators/Templates' => base_path('stubs/generator')], 'generator-stubs');
         }
     }
 }